adicionando novas funçoes no sistema

+++ Views/Users.php
@@ -1,6 +1,7 @@
 <h1>Usuários</h1>
-
+         <?php if($user_edit): ?>
         <a href="<?php echo BASE_URL; ?>Users/add"><div class="button">Adicionar Usuário</div></a>
+        <?php endif; ?>
         <table width="100%" border="0">
             <tr>
                 <th>Email</th>
@@ -14,12 +15,14 @@
                     <td><?php echo $user['email']; ?></td>
                     <td><?php echo $user['name']; ?></td>
                     <td width="160">
+                    <?php if($user_edit): ?>
                         <a href="<?php echo BASE_URL; ?>Users/edit/<?php echo $user['id']; ?>">
-                            <div class="button button_small">Editar</div></a>
+                        <div class="button button_small">Editar</div></a>
 
                         <a href="<?php echo BASE_URL; ?>Users/delete/<?php echo $user['id']; ?>" 
-                           onclick="return confirm('Tem certeza que deseja excluir ?')">
-                            <div class="button button_small">Excluir</div></a>
+                        onclick="return confirm('Tem certeza que deseja excluir ?')">
+                        <div class="button button_small">Excluir</div></a>
+                    <?php endif; ?>
                     </td>
                 </tr>
